refactor: 기존 패키지 구조 변경
